Added Residence-index on 'GetResidences' and empty edit-component. Changed residence-model to use decimal where API does. Added a nav-item tp residence-index.

+++ FribergBlazorApp/Models/Residence.cs
@@ -7,14 +7,14 @@
         public Category category { get; set; }
         public string address { get; set; }
         public Municipality municipality { get; set; }
-        public int startingPrice { get; set; }
-        public int livingArea { get; set; }
-        public int biArea { get; set; }
-        public int plotArea { get; set; }
+        public decimal startingPrice { get; set; }
+        public decimal livingArea { get; set; }
+        public decimal biArea { get; set; }
+        public decimal plotArea { get; set; }
         public string objectDescription { get; set; }
         public int numberOfRooms { get; set; }
-        public int monthlyFee { get; set; }
-        public int operatingCostPerYear { get; set; }
+        public decimal monthlyFee { get; set; }
+        public decimal operatingCostPerYear { get; set; }
         public int constructionYear { get; set; }
         public Picture[] pictures { get; set; }
         public Realtor realtor { get; set; }
